Responses/requirements.txt
sounddevice==0.4.6
vosk==0.3.45
pyttsx3==2.90
faiss-cpu==1.7.4
sentence-transformers==2.2.2
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
bleak==0.21.1